medicalEquipment GET working
