--- backEnd/Routes/postRoutes.js
@@ -1,21 +1,0 @@
-let express = require("express");
-let router = express.Router();
-let Controllers = require("../controllers"); //index.js
-
-router.get("/", (req, res) => {
-  Controllers.postController.getPosts(res);
-});
-
-router.post("/create", (req, res) => {
-  Controllers.postController.createPost(req.body, res);
-});
-
-router.put("/:id", (req, res) => {
-  Controllers.postController.updatePost(req, res);
-});
-
-router.delete("/:id", (req, res) => {
-  Controllers.postController.deletePost(req, res);
-});
-
-module.exports = router;
