+++ backEnd/controllers/index.js
@@ -1,7 +1,6 @@
 module.exports = {
   userController: require("./userController"),
-  postController: require("./postController"),
+  medicalEquipmentController: require("./medicalEquipmentController"),
+
   likeController: require("./likeController"),
 };
-
-// controller.userController.updateUser();
