+++ backEnd/models/index.js
@@ -1,22 +1,22 @@
 "use strict";
 
 const User = require("./users");
+const MedicalEquipment = require("./medicalEquipment");
 
 // const Assessment = require("./assessments");
-// const MedicalEquipment = require("./medicalEquipment");
 
 async function init() {
   await User.sync();
+  await MedicalEquipment.sync();
 
   // await Assessment.sync();
-  // await MedicalEquipment.sync();
 }
 
 init();
 
 module.exports = {
   User, // Exporta el modelo User
+  MedicalEquipment,
 
   // Assessment,
-  // MedicalEquipment,
 };
