+++ backEnd/server.js
@@ -14,14 +14,14 @@ app.get("/", (req, res) => {
 //this is to import routes
 
 const userRoutes = require("./Routes/userRoutes");
+const medicalEquipmentRoutes = require("./Routes/medicalEquipmentRoutes");
 
-// const medicalEquipmentRoutes = require("./Routes/medicalEquipmentRoutes");
 // const assessmentRoutes = require("./Routes/assessmentRoutes");
 
 //this are server Routes
 app.use("/api/users", userRoutes);
+app.use("/api/medical-equipment", medicalEquipmentRoutes);
 
-// app.use("/api/medical-equipment", medicalEquipmentRoutes);
 // app.use("/api/assessments", assessmentRoutes);
 
 // set port, listen for requests
